+++ monitoring/prometheus.yml
@@ -1,0 +1,29 @@
+global:
+  scrape_interval: 15s
+  evaluation_interval: 15s
+
+rule_files:
+  # - "first_rules.yml"
+  # - "second_rules.yml"
+
+scrape_configs:
+  - job_name: 'prometheus'
+    static_configs:
+      - targets: ['localhost:9090']
+
+  - job_name: 'scheduler'
+    static_configs:
+      - targets: ['scheduler:8081']
+    scrape_interval: 5s
+    metrics_path: /metrics
+
+  - job_name: 'subscriber'
+    static_configs:
+      - targets: ['subscriber:8082']
+    scrape_interval: 5s
+    metrics_path: /metrics
+
+  - job_name: 'redis'
+    static_configs:
+      - targets: ['redis:6379']
+    scrape_interval: 15s
